Implement implicit skip behavior
By default, skipping is explicit, meaning that all pages are included in
the RSS feed, unless explicitly set to be skipped. This commit allows the
user to flip this behavior, so that skipping is implicit: all pages are
skipped, unless set to _not_ be skipped.

This is useful for sites with many pages that do not need to be in the
RSS feed, where it becomes inconvenient to add the skip frontmatter
header to each page.